Updated examples for Agents as Actors and Workflows
