--- cookbook/workflows/multi_agent_lotr/components/agentstate.yaml
@@ -1,16 +1,0 @@
-apiVersion: dapr.io/v1alpha1
-kind: Component
-metadata:
-  name: agentstatestore
-spec:
-  type: state.redis
-  version: v1
-  metadata:
-  - name: redisHost
-    value: localhost:6379
-  - name: redisPassword
-    value: ""
-  - name: keyPrefix
-    value: none
-  - name: actorStateStore
-    value: "true"
