--- cookbook/workflows/multi_agent_lotr/services/wizard/app.py
@@ -1,40 +1,0 @@
-from floki import Agent, AgentService
-from dotenv import load_dotenv
-import asyncio
-import logging
-
-async def main():
-    try:
-        # Define Agent
-        wizard_agent = Agent(
-            role="Wizard",
-            name="Gandalf",
-            goal="Guide the Fellowship with wisdom and strategy, using magic and insight to ensure the downfall of Sauron.",
-            instructions=[
-                "Speak like Gandalf, with wisdom, patience, and a touch of mystery.",
-                "Provide strategic counsel, always considering the long-term consequences of actions.",
-                "Use magic sparingly, applying it when necessary to guide or protect.",
-                "Encourage allies to find strength within themselves rather than relying solely on your power.",
-                "Respond concisely, accurately, and relevantly, ensuring clarity and strict alignment with the task."
-            ]
-        )
-
-        # Expose Agent as an Actor over a Service
-        wizard_service = AgentService(
-            agent=wizard_agent,
-            message_bus_name="messagepubsub",
-            agents_state_store_name="agentstatestore",
-            port=8002,
-            daprGrpcPort=50002
-        )
-
-        await wizard_service.start()
-    except Exception as e:
-        print(f"Error starting service: {e}")
-
-if __name__ == "__main__":
-    load_dotenv()
-
-    logging.basicConfig(level=logging.INFO)
-    
-    asyncio.run(main())
