+++ cookbook/workflows/multi_agents/basic_lotr_agents_as_actors/components/agentstate.yaml
@@ -1,0 +1,16 @@
+apiVersion: dapr.io/v1alpha1
+kind: Component
+metadata:
+  name: agentsregistrystore
+spec:
+  type: state.redis
+  version: v1
+  metadata:
+  - name: redisHost
+    value: localhost:6379
+  - name: redisPassword
+    value: ""
+  - name: keyPrefix
+    value: none
+  - name: actorStateStore
+    value: "true"
