+++ cookbook/workflows/multi_agents/basic_lotr_agents_as_actors/components/pubsub.yaml
