+++ cookbook/workflows/multi_agents/basic_lotr_agents_as_actors/dapr.yaml
@@ -1,0 +1,51 @@
+# https://docs.dapr.io/developing-applications/local-development/multi-app-dapr-run/multi-app-template/#template-properties
+version: 1
+common:
+  resourcesPath: ./components
+  logLevel: info
+  appLogDestination: console
+  daprdLogDestination: console
+  configFilePath: config.yaml
+
+apps:
+- appId: HobbitApp
+  appDirPath: ./services/hobbit/
+  appPort: 8001
+  command: ["python3", "app.py"]
+  daprGRPCPort: 50001
+
+- appId: WizardApp
+  appDirPath: ./services/wizard/
+  appPort: 8002
+  command: ["python3", "app.py"]
+  daprGRPCPort: 50002
+
+- appId: ElfApp
+  appDirPath: ./services/elf/
+  appPort: 8003
+  command: ["python3", "app.py"]
+  daprGRPCPort: 50003
+
+- appId: DwarfApp
+  appDirPath: ./services/dwarf/
+  appPort: 8004
+  command: ["python3", "app.py"]
+  daprGRPCPort: 50004
+
+- appId: RangerApp
+  appDirPath: ./services/ranger/
+  appPort: 8007
+  command: ["python3", "app.py"]
+  daprGRPCPort: 50007
+
+- appId: EagleApp
+  appDirPath: ./services/eagle/
+  appPort: 8008
+  command: ["python3", "app.py"]
+  daprGRPCPort: 50008
+
+- appId: WorkflowApp
+  appDirPath: ./services/workflow-llm/
+  appPort: 8009
+  command: ["python3", "app.py"]
+  daprGRPCPort: 50009
