+++ cookbook/workflows/multi_agents/basic_lotr_agents_as_actors/services/ranger/app.py
