+++ cookbook/workflows/multi_agents/basic_lotr_agents_as_actors/services/workflow-random/app.py
@@ -1,0 +1,28 @@
+from floki import RandomOrchestrator
+from dotenv import load_dotenv
+import asyncio
+import logging
+
+async def main():
+    try:
+        random_workflow_service = RandomOrchestrator(
+            name="Orchestrator",
+            essage_bus_name="messagepubsub",
+            state_store_name="agenticworkflowstate",
+            state_key="workflow_state",
+            agents_registry_store_name="agentsregistrystore",
+            agents_registry_key="agents_registry",
+            service_port=8009,
+            daprGrpcPort=50009,
+            max_iterations=3
+        )
+        await random_workflow_service.start()
+    except Exception as e:
+        print(f"Error starting service: {e}")
+
+if __name__ == "__main__":
+    load_dotenv()
+
+    logging.basicConfig(level=logging.INFO)
+    
+    asyncio.run(main())
