+++ cookbook/workflows/multi_agents/basic_lotr_agents_as_workflows/config.yaml
@@ -1,0 +1,12 @@
+apiVersion: dapr.io/v1alpha1
+kind: Configuration
+metadata:
+  name: mydaprconfig
+spec:
+  tracing:
+    samplingRate: "1"
+    zipkin:
+      endpointAddress: http://localhost:9411/api/v2/spans
+  features:
+    - name: SchedulerReminders
+      enabled: false
