+++ cookbook/workflows/multi_agents/basic_lotr_agents_as_workflows/services/hobbit/Frodo_state.json
@@ -1,0 +1,1 @@
+{"instances": {}}
